fix: extract platform-agnostic OpenCode types to fix web build
opencode-runtime.ts uses Bun.spawn which breaks the web tsc build
when reached via type-only import chains. Extract shared types and
helpers into opencode-types.ts and redirect web-reachable imports
there. Add declare for Bun global in opencode-runtime.ts as a
fallback for transitive resolution.
